[Test] 스쿼드/프로젝트/멤버/기술스택 단위 테스트 작성

+++ src/main/java/com/nexus/sion/feature/project/command/application/service/ProjectCommandServiceImpl.java
@@ -210,8 +210,10 @@ public class ProjectCommandServiceImpl implements ProjectCommandService {
     ProjectFpSummary fpSummary =
         projectFpSummaryRepository.findByProjectCode(projectId).orElse(null);
 
-    projectFunctionEstimateRepository.deleteByProjectFpSummaryId(fpSummary.getId());
-    projectFpSummaryRepository.deleteByProjectCode(projectId);
+    if (fpSummary != null) {
+      projectFunctionEstimateRepository.deleteByProjectFpSummaryId(fpSummary.getId());
+      projectFpSummaryRepository.deleteByProjectCode(projectId);
+    }
 
     Project project =
         projectRepository
